agrego todo double

+++ 02-Interfaces/Conversion.c
@@ -1,9 +1,9 @@
 #include "Conversion.h"
 
-double Celsius (int fahr){ // recibe fahrenheit, devuelve celsius
+double Celsius (double fahr){ // recibe fahrenheit, devuelve celsius
 	return ((5.0/9.0) * (fahr-32.0));
 }
 
-double Fahrenheit (int cel){ // recibe celsius, devuelve fahrenheit
+double Fahrenheit (double cel){ // recibe celsius, devuelve fahrenheit
 	return ((9.0/5.0)*cel+32);
 }
